+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Changed
 
  - Redesign machine name
+ - Redesign settings page
 
 ## [0.0.3]